move hash interfaces
