+++ docs/hazmat/primitives/mac/hmac.rst
@@ -18,7 +18,7 @@ of a message.
 .. class:: HMAC(key, algorithm, backend)
 
     HMAC objects take a ``key`` and a
-    :class:`~cryptography.hazmat.primitives.interfaces.HashAlgorithm` provider.
+    :class:`~cryptography.hazmat.primitives.hashes.HashAlgorithm` provider.
     The ``key`` should be :doc:`randomly generated bytes </random-numbers>` and
     is recommended to be equal in length to the ``digest_size`` of the hash
     function chosen. You must keep the ``key`` secret.
@@ -39,7 +39,7 @@ of a message.
     raised.
 
     If ``algorithm`` isn't a
-    :class:`~cryptography.hazmat.primitives.interfaces.HashAlgorithm` provider
+    :class:`~cryptography.hazmat.primitives.hashes.HashAlgorithm` provider
     then ``TypeError`` will be raised.
 
     To check that a given signature is correct use the :meth:`verify` method.
@@ -56,7 +56,7 @@ of a message.
 
     :param bytes key: Secret key as ``bytes``.
     :param algorithm: An
-        :class:`~cryptography.hazmat.primitives.interfaces.HashAlgorithm`
+        :class:`~cryptography.hazmat.primitives.hashes.HashAlgorithm`
         provider such as those described in
         :ref:`Cryptographic Hashes <cryptographic-hash-algorithms>`.
     :param backend: An
